# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jewelry Store REST API",
    "description": "REST API theo database.sql 25 bảng. Đặt hàng COD; phân trang bắt đầu từ 0. KHACH_HANG, NHAN_VIEN, QUAN_LY. Xem README/API.md để biết phân quyền và luồng nghiệp vụ.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/cart/items/{id}": {
      "put": {
        "tags": [
          "cart-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: KHACH_HANG. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "update_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuantityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "KHACH_HANG"
        ]
      },
      "delete": {
        "tags": [
          "cart-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: KHACH_HANG. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "remove_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "KHACH_HANG"
        ]
      }
    },
    "/api/brands/{id}": {
      "get": {
        "tags": [
          "brand-controller"
        ],
        "description": "Public. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "get_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBrandResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [],
        "x-roles": []
      },
      "put": {
        "tags": [
          "brand-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "update_5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBrandResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "QUAN_LY"
        ]
      },
      "delete": {
        "tags": [
          "brand-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "delete_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "QUAN_LY"
        ]
      }
    },
...
    "/api/cart/items": {
      "post": {
        "tags": [
          "cart-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: KHACH_HANG. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "add_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "KHACH_HANG"
        ]
      }
    },
    "/api/brands": {
      "get": {
        "tags": [
          "brand-controller"
        ],
        "description": "Public. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "list_4",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListBrandResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [],
        "x-roles": []
      },
      "post": {
        "tags": [
          "brand-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "create_5",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBrandResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "QUAN_LY"
        ]
      }
    },
...
    "/api/cart/items/{id}/selection": {
      "patch": {
        "tags": [
          "cart-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: KHACH_HANG. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "select",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "KHACH_HANG"
        ]
      }
    },
...
    "/api/cart": {
      "get": {
        "tags": [
          "cart-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: KHACH_HANG. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "get_6",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "KHACH_HANG"
        ]
      },
      "delete": {
        "tags": [
          "cart-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: KHACH_HANG. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "clear",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCartResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "KHACH_HANG"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {}
        }
      },
...
      "QuantityRequest": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "integer",
            "format": "int32",
            "maximum": 1000,
            "minimum": 1
          }
        }
      },
      "ApiResponseCartResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/CartResponse"
          }
        }
      },
      "CartItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "variantId": {
            "type": "integer",
            "format": "int64"
          },
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "productName": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "material": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "unitPrice": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "available": {
            "type": "integer",
            "format": "int32"
          },
          "selected": {
            "type": "boolean"
          }
        }
      },
      "CartResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItemResponse"
            }
          },
          "total": {
            "type": "number"
          }
        }
      },
      "BrandRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "logo": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "HOAT_DONG",
              "NGUNG_HOAT_DONG"
            ]
          }
        },
        "required": [
          "name"
        ]
      },
      "ApiResponseBrandResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/BrandResponse"
          }
        }
      },
      "BrandResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "logo": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "HOAT_DONG",
              "NGUNG_HOAT_DONG"
            ]
          }
        }
      },
...
      "CartItemRequest": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "integer",
            "format": "int64",
            "exclusiveMinimum": 0
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "maximum": 1000,
            "minimum": 1
          }
        },
        "required": [
          "variantId"
        ]
      },
...
      "SelectionRequest": {
        "type": "object",
        "properties": {
          "selected": {
            "type": "boolean"
          }
        },
        "required": [
          "selected"
        ]
      },
...
      "ApiResponseListBrandResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrandResponse"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}